+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 
 - Renamed `converters.rs` file in Rust to `mixins.rs` to better reflect its content and purpose.
 
+### 🔄 Refactors
+
+- **Internal**: Renamed `PyOk` and `PyErr` to `PyoOk` and `PyoErr` to avoid any confusion with `Pyo3` types.
+
 ### 📖 Documentation
 
 - Various reorganisation changes to reduce redundancy